fetch 함수 분리 추후 í
