+++ src/components/indexPage.js
@@ -1,33 +1,11 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import Header from './header.js';
 import Footer from './footer.js';
-import searchIcon from '../assets/images/index/Search.svg';
-import { getFolderData } from '../services/folderService';
 import MainCard from './mainCard';
+import useFolderState from './useFolderState';
+import searchIcon from '../assets/images/index/Search.svg';
 function IndexPage() {
-  const [folder, setFolder] = useState({
-    folder: {
-      links: [],
-      owner: {
-        profileImageSource: '',
-        name: '',
-      },
-      name: '',
-    },
-  });
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const data = await getFolderData();
-        setFolder(data);
-      } catch (error) {
-        console.error('폴더 데이터를 불러오는데 실패했습니다', error);
-      }
-    };
-
-    fetchData();
-  }, []);
+  const { folder } = useFolderState();
 
   return (
     <div>
